+++ src/Traits/AdminMethodsEdit.php
@@ -3,7 +3,6 @@
 namespace Larrock\Core\Traits;
 
 use View;
-use Breadcrumbs;
 use JsValidator;
 use Larrock\Core\Component;
 
@@ -16,7 +15,6 @@ trait AdminMethodsEdit
      * Show the form for editing the specified resource.
      * @param  int $id
      * @return View
-     * @throws \DaveJamesMiller\Breadcrumbs\Exceptions\DuplicateBreadcrumbException
      */
     public function edit($id)
     {
@@ -25,48 +23,6 @@ trait AdminMethodsEdit
 
         $validator = JsValidator::make(Component::_valid_construct($this->config, 'update', $id));
         View::share('validator', $validator);
-
-        Breadcrumbs::register('admin.'.$this->config->name.'.edit', function ($breadcrumbs, $data) {
-            $current_level = null;
-            $breadcrumbs->parent('admin.'.$this->config->name.'.index');
-            if ($data->getCategory) {
-                if (isset($data->getCategory->id)) {
-                    foreach ($data->getCategory->parent_tree as $item) {
-                        $active = ' [Не опубликован!]';
-                        if ($item->active === 1) {
-                            $active = '';
-                        }
-                        $breadcrumbs->push($item->title.$active, '/admin/'.$this->config->name.'/'.$item->id);
-                    }
-                    $current_level = $this->config->getModel()->whereCategory($data->getCategory->id)->orderBy('updated_at', 'DESC')->take('15')->get();
-                } else {
-                    if (\count($data->getCategory) > 0) {
-                        foreach ($data->getCategory->first()->parent_tree as $item) {
-                            $active = ' [Не опубликован!]';
-                            if ($item->active === 1) {
-                                $active = '';
-                            }
-                            $breadcrumbs->push($item->title.$active, '/admin/'.$this->config->name.'/'.$item->id);
-                        }
-                    } else {
-                        $breadcrumbs->push('[Раздел не найден]');
-                    }
-                }
-            } else {
-                if ($data->parent) {
-                    $breadcrumbs->push($data->getConfig()->title, '/admin/'.$data->getConfig()->name.'/'.$data->parent);
-                }
-            }
-            if ($data->title) {
-                $active = ' [Не опубликован!]';
-                if ($data->active === 1) {
-                    $active = '';
-                }
-                $breadcrumbs->push($data->title.$active, '/admin/'.$this->config->getName().'/'.$data->id, ['current_level' => $current_level]);
-            } else {
-                $breadcrumbs->push('Элемент');
-            }
-        });
 
         return view('larrock::admin.admin-builder.edit', $data);
     }