+++ views/admin/admin-builder/edit.blade.php
@@ -5,7 +5,7 @@
     <div class="container-head uk-margin-bottom">
         <div class="uk-grid uk-grid-small">
             <div class="uk-width-expand">
-                {!! Breadcrumbs::render('admin.'. $app->name .'.edit', $data) !!}
+                {!! Breadcrumbs::render('admin.edit', $data) !!}
                 <a class="link-blank uk-display-block" href="{{ $data->full_url }}/">{{ $data->full_url }}/</a>
                 @if(isset($data->getCategory) && count($data->getCategory) > 1)
                     <div>В разделах:</div>